minor clarifications, and add a "summary" section
Feedback from operators is that some vendors ignore the normative
text.  Instead, they rely on misinterpretations of history and other
descriptions in order to behave in weird manners.

As such, we clarify the problematic text, and add a new section
which explains in short form, for people are are confused by pages
of normative content.